requirements.txt
python-dotenv
pandas
twarc
seaborn